Added new features
Added:
Automatic Eating (bread, or a chosen item in the variable)
Block Collection: Finds nearest block that is inputted, paths to it, mines it, and tries to pick it up.
General Utility functions
Reequip last item after eating
Equip function returns true if it found and equipped the item.
Food  amount to the GUI
Block Collection in GUI